환전

+++ flobank-api/src/main/java/kr/co/api/flobankapi/dto/CustTranHistDTO.java
@@ -14,7 +14,7 @@ public class CustTranHistDTO {
     private String tranAcctNo;
     private String tranCustName;
     private Integer tranType;
-    private Integer tranAmount;
+    private BigDecimal tranAmount;
     private String tranDt;
     private String tranRecAcctNo;
     private String tranRecName;
