+++ flobank-api/src/main/java/kr/co/api/flobankapi/dto/FrgnRemtTranDTO.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -11,14 +13,14 @@ public class FrgnRemtTranDTO {
     long remtNo;
     String remtAcctNo;
     String remtCustName;
-    double remtAmount;
+    BigDecimal remtAmount;
     String remtDt;
     String remtRecAccNo;
     String remtRecName;
     String remtRecBkCode;
     String remtCurrency;
-    double remtAppliedRate;
-    double remtFee;
+    BigDecimal remtAppliedRate;
+    BigDecimal remtFee;
     int remtStatus;
     String remtAddr;
     String remtCity;
