+++ flobank-api/src/main/java/kr/co/api/flobankapi/mapper/RemitMapper.java
@@ -5,10 +5,12 @@ import kr.co.api.flobankapi.dto.FrgnAcctBalanceDTO;
 import kr.co.api.flobankapi.dto.FrgnRemtTranDTO;
 import org.apache.ibatis.annotations.Mapper;
 
+import java.math.BigDecimal;
+
 @Mapper
 public interface RemitMapper {
     void insertFrgnRemtTran(FrgnRemtTranDTO frgnRemtTranDTO);
 
-    void updateMinusBal(double amount, String acctNo);
+    void updateMinusBal(BigDecimal amount, String acctNo);
     CustFrgnAcctDTO selectParAcctNo(String balNo);
 }
